Fix hold-to-generate not firing; move token indicator below the whole input box.
The hold gesture lived on a GestureDetector wrapping an interactive
IconButton -- onTapUp only fires if that recognizer wins the gesture
arena, and the nested button's own tap target was winning instead,
so onTapUp (and therefore the entire hold-timing logic) never ran.
Switch to Listener (onPointerDown/Up/Cancel), which reports raw
pointer events unconditionally regardless of what else claims the
gesture, and drop the now-redundant IconButton in favor of a plain
Center+Icon since it no longer needs its own tap handling.

Move the token usage ring back out of the input bar's internal layout
entirely, into its own thin row below the whole rounded input
container (right-aligned, 16px tall) instead of between the textbox
and the button row.
